REF: move test matrices into the main package.
